requirements.txt
absl-py==0.8.1
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ConfigArgParse==0.15.1
Cython==0.29.14
future==0.18.2
google-auth==1.7.2
google-auth-oauthlib==0.4.1
grpcio==1.25.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Markdown==3.1.1
MarkupSafe==1.1.1
nltk==3.4.5
numpy==1.17.4
oauthlib==3.1.0
OpenNMT-py==1.0.0rc2
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyonmttok==1.17.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
sacremoses==0.0.35
six==1.13.0
subword-nmt==0.3.7
tensorboard==2.1.0
torch==1.5.0
torchtext==0.4.0
tqdm==4.30.0
urllib3==1.25.7
Werkzeug==0.16.0
transformers==2.5.1
pyyaml==6.0
flask
flask_json
